# nerava-backend-v9 2/app/openapi-actions.yaml
openapi: 3.1.0
info:
  title: Nerava Actions
  version: 1.0.0
paths:
  /v1/discover:
    post:
      operationId: discover
      x-openai-isConsequential: false
      summary: Unified discovery of nearby events and charging options
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                lat:
                  type: number
                lng:
                  type: number
                radius_m:
                  type: integer
                  default: 1500
                  minimum: 100
                  maximum: 10000
                categories:
                  type: array
                  items:
                    type: string
                    enum: [ev_charging, event, merchant]
                limit:
                  type: integer
                  default: 25
                  minimum: 1
                  maximum: 50
                cursor:
                  type: string
                  nullable: true
                fields:
                  type: array
                  nullable: true
                  items:
                    type: string
                    enum: [id, kind, title, name, category, lat, lng, distance_m, starts_at, ends_at, green_window, offer, cta]
              required: [lat, lng]
      responses:
        '200':
          description: Paginated discovery results
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  items:
                    type: array
                    maxItems: 50
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        kind: { type: string, enum: [event, ev_charging, merchant] }
                        title: { type: string, nullable: true }
                        name: { type: string, nullable: true }
                        category: { type: string, nullable: true }
                        lat: { type: number }
                        lng: { type: number }
                        distance_m: { type: number }
                        starts_at: { type: string, format: date-time, nullable: true }
                        ends_at: { type: string, format: date-time, nullable: true }
                        green_window:
                          type: object
                          nullable: true
                          properties:
                            start: { type: string }
                            end: { type: string }
                        offer:
                          type: object
                          nullable: true
                          properties:
                            title: { type: string }
                            est_reward_cents: { type: integer }
                        cta:
                          type: object
                          nullable: true
                          properties:
                            join_event_id: { type: string, nullable: true }
                            verify_url: { type: string, nullable: true }
                  next_cursor: { type: string, nullable: true }
                  total_estimate: { type: integer, nullable: true }
  /v1/affiliate/track_click:
    post:
      operationId: track_click
      summary: Track affiliate click
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackClickReq'
      responses:
        '200':
          description: OK
  /v1/affiliate/notify:
    post:
      operationId: affiliate_notify
      summary: Ingest affiliate conversion notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AffiliateNotifyReq'
      responses:
        '200':
          description: OK
  /v1/insights/events:
    post:
      operationId: events_insights
      summary: Event insights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsightsReq'
      responses:
        '200':
          description: OK
  /v1/insights/merchants:
    post:
      operationId: merchants_insights
      summary: Merchant insights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsightsReq'
      responses:
        '200':
          description: OK
  /v1/sessions/verify/start:
    post:
      operationId: verify_start
      summary: Start verify dwell session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, lat, lng, accuracy_m, ua]
              properties:
                token:
                  type: string
                lat:
                  type: number
                lng:
                  type: number
                accuracy_m:
                  type: number
                ua:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  session_id:
                    type: string
                  target:
                    type: object
                    nullable: true
                  reason:
                    type: string
                    nullable: true
                  hint:
                    type: object
                    nullable: true
  /v1/sessions/verify/ping:
    post:
      operationId: verify_ping
      summary: Ping verify dwell session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, lat, lng, accuracy_m]
              properties:
                session_id:
                  type: string
                lat:
                  type: number
                lng:
                  type: number
                accuracy_m:
                  type: number
                ts:
                  type: number
                  nullable: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  verified:
                    type: boolean
                  rewarded:
                    type: boolean
                    nullable: true
                  reward_cents:
                    type: integer
                    nullable: true
                  wallet_delta_cents:
                    type: integer
                    nullable: true
                  pool_delta_cents:
                    type: integer
                    nullable: true
                  dwell_seconds:
                    type: integer
                  needed_seconds:
                    type: integer
                    nullable: true
                  distance_m:
                    type: number
                    nullable: true
                  target_acquired:
                    type: boolean
                    nullable: true
                  reason:
                    type: string
                    nullable: true
  /v1/pool/summary:
    get:
      operationId: pool_summary
      summary: Pool summary
      parameters:
        - in: query
          name: city
          schema:
            type: string
          required: false
        - in: query
          name: range
          schema:
            type: string
            enum: [today, 7d, 30d]
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [balance_cents, inflows, outflows, impact]
                properties:
                  city:
                    type: string
                    nullable: true
                  range:
                    type: string
                  balance_cents:
                    type: integer
                  inflows:
                    type: object
                    additionalProperties:
                      type: integer
                  outflows:
                    type: object
                    additionalProperties:
                      type: integer
                  impact:
                    type: object
                    properties:
                      verified_sessions:
                        type: integer
                      avg_reward_cents:
                        type: integer
                  updated_at:
                    type: string
                    format: date-time
components:
  schemas:
    DiscoverReq:
      type: object
      properties:
        user_id:
          type: integer
          nullable: true
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        radius_m:
          type: integer
          default: 2000
        categories:
          type: array
          items:
            type: string
          nullable: true
        time:
          type: string
          format: date-time
          nullable: true
      required: [lat, lng]
    DiscoverRes:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DiscoverItem'
    DiscoverItem:
      type: object
      properties:
        kind:
          type: string
          enum: [event, merchant]
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        distance_m:
          type: number
        start_time:
          type: string
          format: date-time
          nullable: true
        end_time:
          type: string
          format: date-time
          nullable: true
        offer:
          type: object
          nullable: true
        ctas:
          type: array
          items:
            type: object
    TrackClickReq:
      type: object
      required: [user_id, merchant_id, offer_id]
      properties:
        user_id:
          type: integer
        merchant_id:
          type: integer
        offer_id:
          type: string
    AffiliateNotifyReq:
      type: object
      required: [network, click_id, amount_cents]
      properties:
        network:
          type: string
        click_id:
          type: string
        amount_cents:
          type: integer
        merchant_ref:
          type: string
          nullable: true
        meta:
          type: object
          nullable: true
    InsightsReq:
      type: object
      properties:
        merchant_id:
          type: integer
          nullable: true
        event_id:
          type: integer
          nullable: true
        from_ts:
          type: string
          format: date-time
          nullable: true
        to_ts:
          type: string
          format: date-time
          nullable: true
openapi: 3.1.0
info:
  title: Nerava Actions API
  version: "2025-10-29"
  description: >
    GPT-first API for Nerava: Find → Verify → Earn → Spend → Share.
    Includes charger & merchant discovery, session verification, events, and community pool.

servers:
  - url: https://debug-weekend-mpeg-handling.trycloudflare.com
    description: Cloudflare tunnel (demo)

security:
  - ApiKeyAuth: []

paths:
  /health:
    get:
      operationId: health
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthRes'

  /v1/gpt/find_charger:
    get:
      operationId: findCharger
      summary: Find EV chargers near a point
      parameters:
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lng
          required: true
          schema: { type: number }
        - in: query
          name: radius_m
          required: false
          schema: { type: integer, default: 3000, minimum: 50, maximum: 20000 }
        - in: query
          name: city
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List of chargers or an empty-state hint
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Charger' }
                  - $ref: '#/components/schemas/HintResponse'

  /v1/gpt/find_merchants:
    get:
      operationId: findMerchants
      summary: Find merchants near a point (optional category)
      parameters:
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lng
          required: true
          schema: { type: number }
        - in: query
          name: category
          required: false
          schema: { type: string, enum: [coffee, gym, qsr, wellness, other] }
        - in: query
          name: radius_m
          required: false
          schema: { type: integer, default: 1200, minimum: 50, maximum: 20000 }
      responses:
        "200":
          description: List of merchants or an empty-state hint
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Merchant' }
                  - $ref: '#/components/schemas/HintResponse'

  /v1/gpt/create_session_link:
    post:
      operationId: createSessionLink
      summary: Create a public verify URL (short-lived JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionLinkReq' }
      responses:
        "200":
          description: Verify link created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateSessionLinkRes' }

  /v1/sessions/locate:
    post:
      operationId: locate
      summary: Complete verification by posting device GPS
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocateReq' }
      responses:
        "200":
          description: Verification result (includes friendly message)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LocateRes' }

  /v1/gpt/me:
    get:
      operationId: me
      summary: User wallet + stats snapshot
      parameters:
        - in: query
          name: user_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: User snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeRes' }

  /v1/events:
    post:
      operationId: createEvent
      summary: Create an event (activator)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventCreateReq' }
      responses:
        "200":
          description: Event created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }

  /v1/events/nearby:
    get:
      operationId: eventsNearby
      summary: List events near a point
      parameters:
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lng
          required: true
          schema: { type: number }
        - in: query
          name: radius_m
          required: false
          schema: { type: integer, default: 2000, minimum: 50, maximum: 20000 }
        - in: query
          name: now
          required: false
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Nearby events (sorted by distance then start time)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EventNearby' }

  /v1/events/{event_id}/join:
    post:
      operationId: joinEvent
      summary: Join an event
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventJoinReq' }
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventJoinRes' }

  /v1/events/{event_id}/verify/start:
    post:
      operationId: startEventVerification
      summary: Start event verification
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationStartReq' }
      responses:
        "200":
          description: Verification started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationStartRes' }

  /v1/events/verify/{verification_id}/complete:
    post:
      operationId: completeEventVerification
      summary: Complete event verification (geo / optional photo/QR)
      parameters:
        - in: path
          name: verification_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationCompleteReq' }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationCompleteRes' }

  /v1/pool/summary:
    get:
      operationId: poolSummary
      summary: City community pool summary
      parameters:
        - in: query
          name: city
          required: true
          schema: { type: string }
        - in: query
          name: range
          required: false
          schema: { type: string, enum: [today, 7d, 30d], default: 7d }
      responses:
        "200":
          description: Pool summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoolSummaryRes' }

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key

  schemas:
    HealthRes:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]

    HintResponse:
      type: object
      properties:
        items:
          type: array
          items: {}
          description: Always empty when hint is present
        hint:
          type: object
          properties:
            suggestions:
              type: array
              items: { type: string }
            next_steps:
              type: string
          required: [suggestions, next_steps]
      required: [items, hint]

    Charger:
      type: object
      properties:
        charger_id: { oneOf: [ { type: string }, { type: integer } ] }
        name: { type: string }
        lat: { type: number }
        lng: { type: number }
        network: { type: string }
        distance_m: { type: number }
        green_window:
          type: object
          properties:
            start: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
            end:   { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
          required: [start, end]
        nearby_merchants:
          type: array
          items:
            type: object
            properties:
              merchant_id: { type: integer }
              name: { type: string }
              category: { type: string }
              distance_m: { type: number }
            required: [merchant_id, name, category, distance_m]
      required: [charger_id, name, lat, lng, network, distance_m, green_window, nearby_merchants]

    Merchant:
      type: object
      properties:
        merchant_id: { type: integer }
        name: { type: string }
        category: { type: string }
        lat: { type: number }
        lng: { type: number }
        distance_m: { type: number }
        has_offer: { type: boolean }
        offer:
          nullable: true
          type: object
          properties:
            title: { type: string }
            window_start: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]:[0-5][0-9]$" }
            window_end:   { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]:[0-5][0-9]$" }
            est_reward_cents: { type: integer }
            source: { type: string, enum: [local, affiliate, square] }
          required: [title, window_start, window_end, est_reward_cents]
      required: [merchant_id, name, category, lat, lng, distance_m, has_offer]

    CreateSessionLinkReq:
      type: object
      properties:
        user_id: { type: integer, minimum: 1 }
        lat: { type: number, nullable: true }
        lng: { type: number, nullable: true }
        charger_hint: { oneOf: [ { type: string }, { type: integer }, { type: 'null' } ] }
      required: [user_id]

    CreateSessionLinkRes:
      type: object
      properties:
        session_id: { oneOf: [ { type: string }, { type: integer } ] }
        url: { type: string }
        expires_at: { type: string, format: date-time }
      required: [session_id, url, expires_at]

    LocateReq:
      type: object
      properties:
        ts: { type: string, format: date-time }
        lat: { type: number }
        lng: { type: number }
        accuracy: { type: number }
        ua: { type: string }
      required: [ts, lat, lng, accuracy]

    LocateRes:
      type: object
      properties:
        verified: { type: boolean }
        status: { type: string }
        reason: { type: string, nullable: true }
        wallet_delta_cents: { type: integer, nullable: true }
        message: { type: string, nullable: true }
      required: [verified, status]

    MeRes:
      type: object
      properties:
        handle: { type: string }
        reputation: { type: integer }
        followers: { type: integer }
        following: { type: integer }
        wallet_cents: { type: integer }
        month_self_cents: { type: integer }
        month_pool_cents: { type: integer }
      required: [handle, reputation, followers, following, wallet_cents, month_self_cents, month_pool_cents]

    # -------- Events Schemas --------

    EventCreate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        city: { type: string }
        lat: { type: number }
        lng: { type: number }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        green_window_start: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
        green_window_end:   { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
        price_cents: { type: integer, minimum: 0 }
        revenue_split:
          type: object
          properties:
            pool_pct: { type: integer, minimum: 0, maximum: 100 }
            activator_pct: { type: integer, minimum: 0, maximum: 100 }
          required: [pool_pct, activator_pct]
        capacity: { type: integer, nullable: true }
        visibility: { type: string, enum: [public, followers, private], default: public }
      required: [title, city, lat, lng, starts_at, ends_at]

    EventCreateReq:
      type: object
      properties:
        user_id: { type: integer, minimum: 1 }
        event: { $ref: '#/components/schemas/EventCreate' }
      required: [user_id, event]

    Event:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        city: { type: string }
        lat: { type: number }
        lng: { type: number }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        green_window:
          type: object
          nullable: true
          properties:
            start: { type: string }
            end:   { type: string }
        price_cents: { type: integer }
        capacity: { type: integer, nullable: true }
        visibility: { type: string }
        status: { type: string }
      required: [id, title, city, lat, lng, starts_at, ends_at, price_cents, visibility, status]

    EventNearby:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        distance_m: { type: number }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        price_cents: { type: integer }
        green_window:
          type: object
          nullable: true
          properties:
            start: { type: string }
            end:   { type: string }
        capacity_left: { type: integer, nullable: true }
      required: [id, title, distance_m, starts_at, ends_at, price_cents]

    EventJoinReq:
      type: object
      properties:
        user_id: { type: integer, minimum: 1 }
      required: [user_id]

    EventJoinRes:
      type: object
      properties:
        attendance_id: { oneOf: [ { type: string }, { type: integer } ] }
        state: { type: string, enum: [invited, joined, checked_in, verified, refunded] }
      required: [attendance_id, state]

    VerificationStartReq:
      type: object
      properties:
        user_id: { type: integer, minimum: 1 }
        mode: { type: string, enum: [geo, qr, photo], default: geo }
        charger_id: { oneOf: [ { type: string }, { type: integer } ], nullable: true }
      required: [user_id, mode]

    VerificationStartRes:
      type: object
      properties:
        verification_id: { type: string }
        status: { type: string, enum: [pending] }
      required: [verification_id, status]

    VerificationCompleteReq:
      type: object
      properties:
        lat: { type: number }
        lng: { type: number }
        photo_url: { type: string, nullable: true }
        qr_code: { type: string, nullable: true }
      required: [lat, lng]

    VerificationCompleteRes:
      type: object
      properties:
        verification_id: { type: string }
        status: { type: string, enum: [passed, failed] }
        reward_cents: { type: integer, nullable: true }
        pool_contribution_cents: { type: integer, nullable: true }
        message: { type: string, nullable: true }
      required: [verification_id, status]

    # -------- Pool Schemas --------

    PoolSummaryRes:
      type: object
      properties:
        balance_cents: { type: integer }
        inflows:
          type: object
          additionalProperties: { type: integer }
        outflows:
          type: object
          additionalProperties: { type: integer }
        impact:
          type: object
          properties:
            verified_sessions: { type: integer }
            avg_reward_cents: { type: integer }
      required: [balance_cents, inflows, outflows, impact]